automation/requirements.txt
web3==6.11.1
python-dotenv==1.0.0
setuptools==69.5.1